+++ hawp/ssl/datasets/synthetic_util.py
@@ -50,7 +50,7 @@ def add_salt_and_pepper(img):
     img[white > 0] = 255
     img[black > 0] = 0
     cv.blur(img, (5, 5), img)
-    return np.empty((0, 2), dtype=np.int)
+    return np.empty((0, 2), dtype=np.int32)
 
 
 def generate_background(size=(960, 1280), nb_blobs=100, min_rad_ratio=0.01,
@@ -189,8 +189,8 @@ def get_line_heatmap(junctions, line_map, size=[480, 640], thickness=1):
         thickness = int(thickness)
 
     # If the junction points are not int => round them and convert to int
-    if not junctions.dtype == np.int:
-        junctions = (np.round(junctions)).astype(np.int)
+    if not junctions.dtype == np.int32:
+        junctions = (np.round(junctions)).astype(np.int32)
 
     # Initialize empty map
     heat_map = np.zeros(size)
@@ -222,8 +222,9 @@ def draw_lines(img, nb_lines=10, min_len=32, min_label_len=32):
     """
     # Set line number and points placeholder
     num_lines = random_state.randint(1, nb_lines)
-    segments = np.empty((0, 4), dtype=np.int)
-    points = np.empty((0, 2), dtype=np.int)
+    segments = np.empty((0, 4), dtype=np.int32)
+    points = np.empty((0, 2), dtype=np.int32)
+    
     background_color = int(np.mean(img))
     min_dim = min(img.shape)
 
@@ -404,11 +405,11 @@ def draw_multiple_polygons(img, max_sides=8, nb_polygons=30, min_len=32,
       max_sides: maximal number of sides + 1
       nb_polygons: maximal number of polygons
     """
-    segments = np.empty((0, 4), dtype=np.int)
-    label_segments = np.empty((0, 4), dtype=np.int)
+    segments = np.empty((0, 4), dtype=np.int32)
+    label_segments = np.empty((0, 4), dtype=np.int32)
     centers = []
     rads = []
-    points = np.empty((0, 2), dtype=np.int)
+    points = np.empty((0, 2), dtype=np.int32)
     background_color = int(np.mean(img))
 
     min_dim = min(img.shape[0], img.shape[1])
@@ -559,8 +560,8 @@ def draw_ellipses(img, nb_ellipses=20):
     Parameters:
       nb_ellipses: maximal number of ellipses
     """
-    centers = np.empty((0, 2), dtype=np.int)
-    rads = np.empty((0, 1), dtype=np.int)
+    centers = np.empty((0, 2), dtype=np.int32)
+    rads = np.empty((0, 1), dtype=np.int32)
     min_dim = min(img.shape[0], img.shape[1]) / 4
     background_color = int(np.mean(img))
     for i in range(nb_ellipses):
@@ -581,7 +582,7 @@ def draw_ellipses(img, nb_ellipses=20):
         col = get_random_color(background_color)
         angle = random_state.rand() * 90
         cv.ellipse(img, (x, y), (ax, ay), angle, 0, 360, col, -1)
-    return np.empty((0, 2), dtype=np.int)
+    return np.empty((0, 2), dtype=np.int32)
 
 
 def draw_star(img, nb_branches=6, min_len=32, min_label_len=64):
@@ -755,7 +756,7 @@ def draw_checkerboard_multiseg(img, max_rows=7, max_cols=7,
 
     label_segments = np.concatenate(label_segments)
     """
-    label_segments = np.empty([0, 4], dtype=np.int)
+    label_segments = np.empty([0, 4], dtype=np.int32)
     # Iterate through rows
     for row_idx in range(rows + 1):
         # Include all the combination of the junctions
@@ -815,7 +816,7 @@ def draw_checkerboard_multiseg(img, max_rows=7, max_cols=7,
         else:
             continue
 
-    label_segments = np.round(label_segments_filtered).astype(np.int)
+    label_segments = np.round(label_segments_filtered).astype(np.int32)
     
     # Only record the segments longer than min_label_len
     points1 = label_segments[:, :2]
@@ -1025,7 +1026,7 @@ def draw_stripes_multiseg(img, max_nb_cols=13, min_len=0.04, min_label_len=64,
         else:
             continue
 
-    segments = (np.round(segments_new)).astype(np.int)
+    segments = (np.round(segments_new)).astype(np.int32)
 
     # Only record the segments longer than min_label_len
     points1 = segments[:, :2]
@@ -1190,7 +1191,7 @@ def draw_cube(img, min_size_ratio=0.2, min_label_len=64,
         else:
             continue
 
-    segments = (np.round(segments_new)).astype(np.int)
+    segments = (np.round(segments_new)).astype(np.int32)
 
     # Only record the segments longer than min_label_len
     points1 = segments[:, :2]
